+++ src/main/java/model/Lotto.java
@@ -1,3 +1,5 @@
+package model;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
