+++ src/main/java/model/WinningNumber.java
@@ -1,7 +1,7 @@
+package model;
+
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 public class WinningNumber {
     private final List<Integer> numbers = new ArrayList<>();
