+++ src/main/java/model/WinningResult.java
@@ -1,5 +1,6 @@
+package model;
+
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
