+++ src/main/java/view/OutputView.java
@@ -1,0 +1,4 @@
+package view;
+
+public class OutputView {
+}
